정규식 체크를 위한 클래스 생성 정규식 체크가 여러 번 필요해 보여서 따로 클래스를 뽑아줌.
