+++ app/src/main/java/com/goodee/cando_app/views/user/FindEmailFragment.kt
@@ -14,12 +14,13 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import com.goodee.cando_app.R
 import com.goodee.cando_app.databinding.FragmentFindIdBinding
+import com.goodee.cando_app.util.RegexChecker
 import com.goodee.cando_app.viewmodel.UserViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class FindIdFragment : Fragment() {
+class FindEmailFragment : Fragment() {
     companion object {
         private const val TAG: String = "로그"
     }
@@ -48,10 +49,10 @@ class FindIdFragment : Fragment() {
             val phone = binding.edittextFindidPhoneinput.text.toString().trim().replace("-","")
             val name = binding.edittextFindidNameinput.text.toString()
 
-            if (phone.isEmpty() || phone.isBlank()) {
-                Toast.makeText(requireActivity(), "이메일을 입력해주세요", Toast.LENGTH_SHORT).show()
-            } else if (name.isEmpty() || name.isBlank()) {
+            if (name.isEmpty() || name.isBlank()) {
                 Toast.makeText(requireActivity(), "이름을 입력해주세요", Toast.LENGTH_SHORT).show()
+            } else if (phone.isEmpty() || phone.isBlank() || !RegexChecker.isValidPhone(phone)) {
+                Toast.makeText(requireActivity(), "전화번호를 확인해주세요", Toast.LENGTH_SHORT).show()
             } else {
                 lifecycleScope.launch(Dispatchers.IO) {
                     val qResult = userViewModel.findUserEmail(phone = phone, name = name)
@@ -63,7 +64,8 @@ class FindIdFragment : Fragment() {
                         withContext(Dispatchers.Main) {
                             // if : 입력한 정보와 일치하는 회원이 여러 명
                             if (qResult.documents.size > 1) {
-                            } else {
+
+                            } else if (qResult.documents.isNotEmpty()){
                                 val alertDialogBuilder = AlertDialog.Builder(requireContext()).create()
                                 alertDialogBuilder.setTitle("찾으시는 이메일")
                                 alertDialogBuilder.setMessage(qResult.documents[0].get("email").toString())
